code cleanup and c error code
