+++ exclusion.dic
@@ -1,2 +1,2 @@
 Qynit
-﻿Biquad
+Biquad
