+++ src/Qynit.PulseGen.Aot/Api.cs
@@ -1,7 +1,6 @@
 using System.Runtime.InteropServices;
 
 using MessagePack;
-using MessagePack.Resolvers;
 
 using Qynit.PulseGen.Aot.Models;
 
@@ -12,24 +11,70 @@ public static class Api
     private static MessagePackSerializerOptions MessagePackSerializerOptions { get; } =
         new MessagePackSerializerOptions(GeneratedMessagePackResolver.InstanceWithStandardAotResolver);
 
-    [UnmanagedCallersOnly(EntryPoint = "Qynit_PulseGen_Run")]
-    public static unsafe IntPtr Run(byte* requestMsg, int requestMsgLen)
+    enum ErrorCode
     {
-        var request = DeserializeRequest(requestMsg, requestMsgLen);
-        var waveforms = GenerateWaveforms(request);
+        Success = 0,
+        DeserializeError = 1,
+        GenerateWaveformsError = 2,
+        KeyNotFound = 3,
+        CopyWaveformError = 4,
+        InvalidHandle = 5,
+        InternalError = 6,
+    }
+
+    [UnmanagedCallersOnly(EntryPoint = "Qynit_PulseGen_Run")]
+    public static unsafe int Run(byte* requestMsg, int requestMsgLen, void** outWaveformDict)
+    {
+        ScheduleRequest request;
+        try
+        {
+            request = DeserializeRequest(requestMsg, requestMsgLen);
+        }
+        catch (Exception e)
+        {
+            Console.Error.WriteLine(e);
+            return (int)ErrorCode.DeserializeError;
+        }
+        List<PooledComplexArray<float>> waveforms;
+        try
+        {
+            waveforms = GenerateWaveforms(request);
+        }
+        catch (Exception e)
+        {
+            Console.Error.WriteLine(e);
+            return (int)ErrorCode.GenerateWaveformsError;
+        }
         var waveformsDict = waveforms.Zip(request.ChannelTable!).ToDictionary(x => x.Second.Name, x => x.First);
         var handle = GCHandle.Alloc(waveformsDict);
-        return GCHandle.ToIntPtr(handle);
+        *outWaveformDict = GCHandle.ToIntPtr(handle).ToPointer();
+        return (int)ErrorCode.Success;
     }
 
     [UnmanagedCallersOnly(EntryPoint = "Qynit_PulseGen_CopyWaveform")]
-    public static unsafe void CopyWaveform(IntPtr handle, IntPtr chName, float* bufferI, float* bufferQ, int bufferLen)
+    public static unsafe int CopyWaveform(IntPtr handle, IntPtr chName, float* bufferI, float* bufferQ, int bufferLen)
     {
+        Dictionary<string, PooledComplexArray<float>> waveformsDict;
         try
         {
-            var waveformsDict = (Dictionary<string, PooledComplexArray<float>>)GCHandle.FromIntPtr(handle).Target!;
-            var chNameStr = Marshal.PtrToStringUTF8(chName);
-            var waveform = waveformsDict[chNameStr];
+            waveformsDict = (Dictionary<string, PooledComplexArray<float>>)GCHandle.FromIntPtr(handle).Target!;
+        }
+        catch (Exception e)
+        {
+            Console.Error.WriteLine(e);
+            return (int)ErrorCode.InvalidHandle;
+        }
+        var chNameStr = Marshal.PtrToStringUTF8(chName);
+        if (chNameStr is null)
+        {
+            return (int)ErrorCode.KeyNotFound;
+        }
+        if (!waveformsDict.TryGetValue(chNameStr, out var waveform))
+        {
+            return (int)ErrorCode.KeyNotFound;
+        }
+        try
+        {
             var spanI = new Span<float>(bufferI, bufferLen);
             waveform.DataI.CopyTo(spanI);
             var spanQ = new Span<float>(bufferQ, bufferLen);
@@ -37,21 +82,31 @@ public static class Api
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
-            throw;
+            Console.Error.WriteLine(e);
+            return (int)ErrorCode.CopyWaveformError;
         }
+        return (int)ErrorCode.Success;
     }
 
     [UnmanagedCallersOnly(EntryPoint = "Qynit_PulseGen_FreeWaveform")]
-    public static unsafe void FreeWaveform(IntPtr handle)
+    public static int FreeWaveform(IntPtr handle)
     {
+        Dictionary<string, PooledComplexArray<float>> waveformsDict;
+        try
+        {
+            waveformsDict = (Dictionary<string, PooledComplexArray<float>>)GCHandle.FromIntPtr(handle).Target!;
+        }
+        catch (Exception e)
+        {
+            Console.Error.WriteLine(e);
+            return (int)ErrorCode.InvalidHandle;
+        }
+        foreach (var waveform in waveformsDict.Values)
+        {
+            waveform.Dispose();
+        }
         GCHandle.FromIntPtr(handle).Free();
-    }
-
-    [UnmanagedCallersOnly(EntryPoint = "Qynit_PulseGen_Hello")]
-    public static void Hello()
-    {
-        throw new Exception("Hello from .NET");
+        return (int)ErrorCode.Success;
     }
 
     private static List<PooledComplexArray<float>> GenerateWaveforms(ScheduleRequest request)
@@ -62,8 +117,11 @@ public static class Api
 
     private static unsafe ScheduleRequest DeserializeRequest(byte* msg, int len)
     {
-        using var stream = new UnmanagedMemoryStream(msg, len);
+        using var memoryManager = new UnsafeMemoryManager<byte>(msg, len);
+        var memory = memoryManager.Memory;
         var options = MessagePackSerializerOptions;
-        return MessagePackSerializer.Deserialize<ScheduleRequest>(stream, options);
+        var formatter = options.Resolver.GetFormatterWithVerify<ScheduleRequest>();
+        var reader = new MessagePackReader(memory);
+        return formatter.Deserialize(ref reader, options);
     }
 }
