+++ src/Qynit.PulseGen.Aot/Models/RepeatElementDto.cs
@@ -1,0 +1,35 @@
+﻿using CommunityToolkit.Diagnostics;
+
+using MessagePack;
+
+using Qynit.PulseGen.Schedules;
+
+namespace Qynit.PulseGen.Aot.Models;
+
+[MessagePackObject]
+public sealed class RepeatElementDto : ScheduleElementDto
+{
+    [Key(6)]
+    public ScheduleElementDto? Element { get; set; }
+    [Key(7)]
+    public int Count { get; set; }
+    [Key(8)]
+    public double Spacing { get; set; }
+
+    public override ScheduleElement GetScheduleElement(ScheduleRequest request)
+    {
+        Guard.IsNotNull(Element);
+        var element = Element.GetScheduleElement(request);
+        return new RepeatElement(element, Count)
+        {
+            Spacing = Spacing,
+            Margin = Margin,
+            Alignment = Alignment,
+            IsVisible = IsVisible,
+            Duration = Duration,
+            MaxDuration = MaxDuration,
+            MinDuration = MinDuration,
+            PulseGenOptions = request.Options?.GetOptions(),
+        };
+    }
+}
