+++ src/Qynit.PulseGen.Aot/Qynit.PulseGen.Aot.csproj
@@ -6,16 +6,15 @@
     <Nullable>enable</Nullable>
     <PublishAot>true</PublishAot>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-      <TrimmerSingleWarn>false</TrimmerSingleWarn>
   </PropertyGroup>
 
-    <ItemGroup>
-        <PackageReference Include="MessagePack" Version="2.6.100-alpha" />
-        <PackageReference Include="MessagePack.SourceGenerator" Version="2.6.100-alpha">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-        </PackageReference>
-    </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="MessagePack" Version="2.6.100-alpha" />
+    <PackageReference Include="MessagePack.SourceGenerator" Version="2.6.100-alpha">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
+    </PackageReference>
+  </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Qynit.PulseGen\Qynit.PulseGen.csproj" />
