+++ src/Qynit.PulseGen.Server/Server.cs
@@ -1,5 +1,3 @@
-using System.Reflection;
-
 using MessagePack;
 using MessagePack.Resolvers;
